backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
sentence-transformers
scikit-learn
numpy
pydantic
openai
google-generativeai
ollama
